@@ -3,32 +3,27 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/react-hooks": "^8.0.1",
-    "@testing-library/user-event": "^12.8.3",
-    "axios": "^0.24.0",
-    "dayjs": "^1.10.7",
+    "axios": "^1.4.0",
+    "dayjs": "^1.11.7",
     "eslint-plugin-prettier": "^4.2.1",
-    "formik": "^2.2.9",
+    "formik": "^2.4.0",
     "fsevents": "^2.3.2",
     "jest-dom": "^4.0.0",
     "jwt-decode": "^3.1.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-hook-form": "^7.22.5",
-    "react-icons": "^4.3.1",
-    "react-router-dom": "^6.2.1",
-    "react-test-renderer": "^17.0.2",
-    "styled-components": "^5.3.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-hook-form": "^7.44.2",
+    "react-icons": "^4.9.0",
+    "react-router-dom": "^6.11.2",
+    "react-test-renderer": "^18.2.0",
+    "styled-components": "^5.3.11",
     "watchman": "^1.0.0",
-    "web-vitals": "^1.1.2",
-    "yup": "^0.32.11"
+    "web-vitals": "^3.3.2",
+    "yup": "^1.2.0"
   },
   "scripts": {
     "start": "react-scripts start --openssl-legacy-provider start",
     "build": "react-scripts build",
-    "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
