@@ -1,12 +1,14 @@
 import React from "react";
-import ReactDOM from "react-dom";
 import App from "./App";
 import GlobalStyle from "./Styles/GlobalStyles";
+import { createRoot } from "react-dom/client";
 
-ReactDOM.render(
+const container = document.getElementById("root");
+const root = createRoot(container);
+
+root.render(
   <React.StrictMode>
     <GlobalStyle />
     <App />
-  </React.StrictMode>,
-  document.getElementById("root")
+  </React.StrictMode>
 );
