@@ -9,11 +9,10 @@ function NewItems() {
   useEffect(() => {
     getAllItems()
       .then((items) => {
-        console.log(items);
         setApiData(items);
       })
       .catch((error) => {
-        console.log(error);
+        console.error(error);
       });
 
     return () => {
