@@ -1,5 +1,5 @@
 import React, { useContext } from "react";
-import { Outlet, Navigate } from "react-router";
+import { Outlet, Navigate } from "react-router-dom";
 import AuthContext from "../Context/AuthContext";
 
 const ProtectedRoute = () => {
